Stripped down to basic webpack app
