+++ package.json
@@ -19,30 +19,12 @@
     "url": "https://github.com/CJACOBSON32/cs4241_Assignment_3/issues"
   },
   "dependencies": {
-    "@types/express-session": "^1.17.5",
-    "@types/passport": "^1.0.11",
-    "@types/passport-local": "^1.0.34",
-    "cookie-session": "^2.0.0",
-    "crypto-js": "^4.1.1",
-    "express": "^4.18.1",
-    "express-session": "^1.17.3",
-    "mongodb": "^4.10.0",
-    "morgan": "^1.10.0",
-    "paper": "^0.12.15",
-    "passport": "^0.6.0",
-    "passport-github2": "^0.1.12",
-    "passport-local": "^1.0.0",
-    "ws": "^8.9.0"
+    "express": "^4.18.1"
   },
   "homepage": "https://github.com/CJACOBSON32/cs4241_Assignment_3#readme",
   "devDependencies": {
-    "@types/cookie-session": "^2.0.44",
-    "@types/crypto-js": "^4.1.1",
     "@types/express": "^4.17.14",
-    "@types/mongodb": "^4.0.7",
-    "@types/morgan": "^1.9.3",
     "@types/node": "^18.7.18",
-    "@types/ws": "^8.5.3",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.3",
     "webpack": "^5.74.0",
