--- src/css/canvas.css
@@ -1,49 +1,0 @@
-
-/*fit to screen*/
-html, body {
-	height: 100vh;
-	max-height: 100vh;
-	overflow: visible;
-}
-
-/*Expand main div to fill screen*/
-#main-sections {
-	display: flex;
-	flex-direction: column;
-	height: 100%;
-	max-height: 100vh;
-	width: 100%;
-	overflow: auto;
-}
-
-/*Canvas*/
-#canvas-section {
-	width: 100%;
-	height: 80%;
-	max-height: 100%;
-	flex-grow: 1;
-}
-
-#canvas-container {
-	padding: 10px;
-	flex-grow: 1;
-	overflow: scroll;
-}
-
-/*Tool Palette*/
-#left-sidebar {
-	width: 200px;
-}
-
-#left-sidebar .mui-container {
-	height: 100%;
-	justify-content: space-between;
-}
-
-.vbox > * {
-	margin: 10px 0;
-}
-
-#left-sidebar label {
-	margin-right: 10px;
-}
