--- src/css/login.css
@@ -1,12 +1,0 @@
-
-body {
-	height: 100vh;
-	width: 100vw;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
-#login-box > h2 {
-	margin-bottom: 20pt;
-}
