+++ src/index.html
@@ -1,18 +1,12 @@
 <!doctype html>
 <html lang="en">
 <head>
-	<title>Inkboard</title>
+	<title>3D Thing</title>
 	<meta charset="utf-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-	<!--	MUI		-->
-	<script src="//cdn.muicss.com/mui-0.10.3/js/mui.min.js"></script>
-	<link href="//cdn.muicss.com/mui-0.10.3/css/mui.min.css" rel="stylesheet" type="text/css" />
-
 	<!--	Styles		-->
 	<link rel="stylesheet" href="css/styles.css" />
-	<link rel="stylesheet" href="css/canvas.css" />
-	<link rel="stylesheet" href="css/material-btn.css" />
 
 	<!--	webpack bundle	-->
 	<script src="dist/index.bundle.js" defer></script>
@@ -20,38 +14,6 @@
 <body>
 	<div id="main-sections">
 
-		<!--	Top Bar	-->
-		<section class="mui-appbar" style="padding-bottom: 20px">
-			<div class="mui-container">
-				<h2>Welcome to Inkboard!</h2>
-				<p>Draw on the canvas and work with others to make art!</p>
-				<p>Click 'clear' to clear all paths that <b>you</b> have made.</p>
-			</div>
-		</section>
-
-		<!--    Canvas    -->
-		<section id="canvas-section" class="hbox">
-			<div class="mui-appbar" id="left-sidebar">
-				<div class="mui-container vbox">
-					<div class="vbox">
-						<button class="mtl-btn" id="clear">Clear</button>
-						<div class="hbox">
-							<label for="color-picker">Stroke Color:</label>
-							<input type="color" id="color-picker">
-						</div>
-						<div class="hbox">
-							<label for="color-picker">Stroke Width:</label>
-							<input type="range" id="width" min="1" max="20" >
-						</div>
-					</div>
-
-					<button class="mtl-btn" id="sign-out" >Sign out</button>
-				</div>
-			</div>
-			<div id="canvas-container">
-				<canvas id="main-canvas" class="mui-panel" height="1950" width="1950"></canvas>
-			</div>
-		</section>
 	</div>
 </body>
 </html>
